+++ little-games/api-layer/src/main/scala/com/github/killjoyer/modules/chat/SimpleChatHandler.scala
@@ -3,7 +3,7 @@ import com.github.killjoyer.domain.chats.Chat.ChatId
 import com.github.killjoyer.domain.users.Username
 import com.github.killjoyer.infrastructure.utils.newtype._
 import com.github.killjoyer.modules.chat.SimpleChatHandler.ChatMessage
-import com.github.killjoyer.services.traits.ChatsManager
+import com.github.killjoyer.services.application.traits.ChatsManager
 import io.circe.generic.JsonCodec
 import zio.UIO
 import zio.ZLayer
