+++ little-games/main/src/test/scala/com/github/killjoyer/services/BullsAndCowsServiceTests.scala
@@ -1,7 +1,7 @@
 package com.github.killjoyer.services
 
 import com.github.killjoyer.mocks.MockDictionaryRepository
-import com.github.killjoyer.services.impls.BullsAndCowsServiceLive
+import com.github.killjoyer.services.domain.impls.BullsAndCowsServiceLive
 import com.github.killjoyer.services.traits.BullsAndCowsService
 import com.github.killjoyer.services.traits.BullsAndCowsService.BullsAndCowsResult
 import zio.Scope
