+++ little-games/service-layer/src/main/scala/com/github/killjoyer/services/application/impls/UserEventsRouterLive.scala
@@ -2,8 +2,8 @@ package com.github.killjoyer.services.impls
 
 import com.github.killjoyer.domain.events.LittleGamesEvent
 import com.github.killjoyer.domain.users.Username
+import com.github.killjoyer.services.application.traits.UserEventsRouter
 import com.github.killjoyer.services.impls.UserEventsRouterLive.UserStorage
-import com.github.killjoyer.services.traits.UserEventsRouter
 import zio.Chunk
 import zio.Hub
 import zio.RIO
