+++ little-games/service-layer/src/main/scala/com/github/killjoyer/services/domain/impls/BullsAndCowsServiceLive.scala
@@ -1,4 +1,4 @@
-package com.github.killjoyer.services.impls
+package com.github.killjoyer.services.domain.impls
 
 import cats.implicits.catsSyntaxEq
