+++ little-games/service-layer/src/main/scala/com/github/killjoyer/services/domain/traits/BullsAndCowsService.scala
